Create BellKAT.Implementation.Configuration and move NetworkCapacity / ExecutionParams related stuff there
