+++ src/BellKAT/Implementations/ProbAtomicOneStepQuantum.hs
@@ -1,10 +1,6 @@
 module BellKAT.Implementations.ProbAtomicOneStepQuantum
     ( ProbAtomicOneStepPolicy
     , ProbAtomicOneStepPolicy'
-    , NetworkCapacity (NC)
-    , ExecutionParams(..)
-    , fromNetworkCapacity
-    , applyExecutionParams
     , execute
     , execute'
     , executeWith
@@ -20,7 +16,6 @@ import qualified Data.Set as Set
 import Data.Default
 import Control.Subcategory.Functor
 
-import qualified Data.Map                            as Map
 import qualified BellKAT.Utils.Multiset              as Mset
 import BellKAT.Utils.Distribution (D', RationalOrDouble, DDom)
 import qualified BellKAT.Utils.Distribution as D
@@ -30,6 +25,7 @@ import BellKAT.Definitions.Tests
 import BellKAT.Definitions.Structures
 import BellKAT.Definitions.Atomic
 import BellKAT.Utils.Choice
+import BellKAT.Implementations.Configuration
 import BellKAT.Implementations.Output
 
 -- | Essentially a symbol for `BellKAT.Utils.Automata.Guarded.GuardedFA` representing a set of
@@ -69,35 +65,6 @@ instance (OpOutput output op tag, Monoid output, Ord output, Ord tag)
                         mempty]
                 else mempty
 
--- | Network capacity, i.e., the maximum number of each possible `BellPair`, essentially
--- a `TaggedBellPairs`
-newtype NetworkCapacity tag = NC 
-    { unNC :: TaggedBellPairs tag 
-    } deriving newtype (Monoid, Semigroup)
-    -- TODO: should really be BellPairs
-
-instance Ord tag => GHC.Exts.IsList (NetworkCapacity tag) where
-    type Item (NetworkCapacity tag) = TaggedBellPair tag
-    fromList = NC . fromList
-    toList = toList . unNC
-
--- | Combine a filter (cut-offs) with network capacity
-data ExecutionParams tag rTag cTag = EP
-    { epNetworkCapacity :: Maybe (NetworkCapacity tag)
-    , epFilter          :: TaggedBellPair rTag -> cTag -> Bool
-    }
-
--- | Construct 'ExecutionParams' from an optional 'NetworkCapacity' and
--- a default permissive filter.
-fromNetworkCapacity :: Maybe (NetworkCapacity tag) -> ExecutionParams tag rTag cTag
-fromNetworkCapacity mbCap = EP { epNetworkCapacity = mbCap
-                               , epFilter          = \_ _ -> True
-                          }
-
-instance Default (ExecutionParams tag rTag cTag) where
-    def = EP { epNetworkCapacity = Nothing
-             , epFilter          = \_ _ -> True
-        }
 
 -- Interprets `ProbAtomicOneStepPolicy` as a monadic function from `TaggedBellPairs` to `CD'` of
 -- `TaggedBellPairs`
@@ -157,41 +124,3 @@ executePAA fix act bps =
          [ cmap (fix . (<> rest)) (computeOutput (paaOutput act) chosen)
            | Partial { chosen , rest }  <- findElemsNDT (fmap staticTag) (toList . paaInputBPs $ act) bps]
         else mempty
-
--- | For each BellPair, keep at most the number allowed
-fixNetworkCapacity 
-    :: (RuntimeTag rTag tag, Ord rTag, Ord tag)
-    => NetworkCapacity tag -> LabelledBellPairs cTag rTag -> LabelledBellPairs cTag rTag
-fixNetworkCapacity (NC cap) (Mset.LMS (ms, t)) =
-  let capCounts = countByBellPair (toList cap)
-      grouped = Map.fromListWith (++) [(staticBellPair tbp, [tbp]) | tbp <- toList ms]
-                                {- ^ duplicates with same key concatenate -}
-      clipped = concat
-        [ take (Map.findWithDefault maxBound bp capCounts) tbps
-        {- ^ keeps only first n tagged instances for that BellPair -}
-        | (bp, tbps) <- Map.toList grouped
-        ]
-  in Mset.LMS (Mset.fromList clipped, t)
-
--- | Count BellPairs in a TaggedBellPairs, disregarding tag
-countByBellPair :: Ord tag => [TaggedBellPair tag] -> Map.Map (TaggedBellPair tag) Int
-countByBellPair xs = Map.fromListWith (+) [ (staticBellPair tbp, 1) | tbp <- xs ]
-
--- | Apply filter and network capacity sequentially
-applyExecutionParams
-    :: (RuntimeTag rTag tag, Ord rTag, Ord tag)
-    => ExecutionParams tag rTag cTag
-    -> LabelledBellPairs cTag rTag
-    -> LabelledBellPairs cTag rTag
-applyExecutionParams (EP mbCap bpPred) bps0 =
-    let afterFilter = fixFilter bpPred bps0
-    in maybe afterFilter (`fixNetworkCapacity` afterFilter) mbCap
-
--- | Filter labelled bell pairs using the given filter (predicate)
-fixFilter :: (Ord rTag)
-               => (TaggedBellPair rTag -> cTag -> Bool)
-               -> LabelledBellPairs cTag rTag
-               -> LabelledBellPairs cTag rTag
-fixFilter bpFilter (Mset.LMS (ms, clock)) =
-    let kept = filter (`bpFilter` clock) (GHC.Exts.toList ms)
-    in Mset.LMS (Mset.fromList kept, clock)
